+++ scalyr_agent/scalyr_client.py
@@ -117,7 +117,7 @@ class ScalyrClientSession(object):
             certificate authorities. This is used for the SSL connections to verify the connection is to Scalyr.
         @param proxies:  A dict describing the network proxies to use (such as a mapping for `https`) or None.
         @param compression_type:  A string containing the compression method to use.
-            Valid options are bz2, deflate or None.  Defaults to None.
+            Valid options are bz2, deflate, lz4, zstandard or None.  Defaults to None.
         @param compression_level: An int containing the compression level of compression to use, from 1-9.  Defaults to 9 (max)
         @param enforce_monotonic_timestamps: A bool that indicates whether event timestamps in the same session
             should be monotonically increasing or not.  Defaults to False
@@ -176,15 +176,19 @@ class ScalyrClientSession(object):
         }
 
         # Configure compression type
+        self.__compression_type = compression_type
+
         self.__compress = None
         encoding = None
 
         if compression_type:
-            if compression_type in ("deflate", "bz2"):
-                compress_func = verify_and_get_compress_func(compression_type)
-                if compress_func:
-                    self.__compress = compress_func
-                    encoding = compression_type
+            compress_func = verify_and_get_compress_func(
+                compression_type, compression_level
+            )
+
+            if compress_func:
+                self.__compress = compress_func
+                encoding = compression_type
 
             if not self.__compress:
                 log.warning(
@@ -196,14 +200,6 @@ class ScalyrClientSession(object):
             self.__standard_headers["Content-Encoding"] = encoding
 
         # Configure compression level
-        if self.__compress:
-            if compression_level < 1 or compression_level > 9:
-                log.warning(
-                    "Invalid compression level used - %d.  Range must be 1-9.  Defaulting to 9 - maximum compression."
-                    % (compression_level)
-                )
-                compression_level = 9
-
         self.__compression_level = compression_level
 
         # The number of sconds to wait for a blocking operation on the connection before considering it to have
@@ -345,7 +341,31 @@ class ScalyrClientSession(object):
             self.total_request_bytes_sent += len(body_str) + len(request_path)
 
             if self.__compress:
-                body_str = self.__compress(body_str, self.__compression_level)
+                size_before_compress = len(body_str)
+
+                start_time = time.time()
+                body_str = self.__compress(body_str)
+                end_time = time.time()
+
+                size_after_compress = len(body_str)
+                compression_ratio = round(
+                    (float(size_before_compress) / size_after_compress), 2
+                )
+                duration = round((end_time - start_time), 4)
+
+                log.log(
+                    scalyr_logging.DEBUG_LEVEL_1,
+                    'Compressed add event request data using "%s" algorithm and level "%s": '
+                    "original_size=%s compressed_size=%s compression_ratio=%s duration=%ss"
+                    % (
+                        self.__compression_type,
+                        self.__compression_level,
+                        size_before_compress,
+                        size_after_compress,
+                        compression_ratio,
+                        duration,
+                    ),
+                )
 
             self.total_compressed_request_bytes_sent += len(body_str) + len(
                 request_path
@@ -488,6 +508,13 @@ class ScalyrClientSession(object):
                         error_code="requestFailed",
                     )
                 return "requestFailed", len(body_str), response
+
+            try:
+                response = six.ensure_text(response, "utf-8", "ignore")
+            except Exception:
+                # We ignore the exception since we still want to log the response even if it
+                # contains non utf-8 characters
+                pass
 
             log.log(
                 scalyr_logging.DEBUG_LEVEL_5,